Decouple agent core from frontend, break circular deps, split settings handler
Three structural improvements to reduce coupling and improve cohesion:

1. Remove @frontend/ imports from agent core (OutputNode, executeAgent):
   - Agent layer now emits typed events via ProgressEventBus instead of
     directly calling UI functions (openBuildDisplayIfTex, showInstruction)
   - New agentEventListeners.ts in frontend registers listeners that bridge
     events to actual UI operations
   - Eliminates the most critical architectural violation: agent flow nodes
     controlling VS Code UI directly

2. Break agent↔tools circular dependency:
   - Move SubagentProgressUpdate types from @tools/subagentResults to
     @shared/schemas/subagentProgress (pure data types, no impl dependency)
   - Agent runtime and tool-use flows now import from @shared/schemas
   - Original file re-exports for backward compatibility
   - Eliminates: executeAgent → @tools/ → executeAgent cycle

3. Split SettingsViewMessageHandler (2041→1087 lines, 47% reduction):
   - Extract AgentHandlers: agent selection, directories, mode presets (~730 lines)
   - Extract LatexSettingsHandlers: tool detection, settings, install (~310 lines)
   - Main handler delegates via typed context interface
   - Each domain is now independently testable and navigable
