+++ src/settingsView/handlers/agentHandlers.ts
@@ -1,0 +1,734 @@
+/**
+ * Agent selection, directory, and mode preset handlers.
+ *
+ * Extracted from SettingsViewMessageHandler to improve cohesion.
+ * Handles agent enable/disable, create/customize/delete, YAML editing,
+ * custom agent directories, and agent mode presets.
+ */
+import * as path from 'path';
+import * as vscode from 'vscode';
+import { z } from 'zod';
+
+import {
+  SETTINGS_VIEW_CMD,
+  type SettingsViewInboundMessage,
+  type AgentSelectionItem,
+} from '@shared/schemas/settingsViewMessages';
+import {
+  AGENT_MODE_PRESETS,
+  AgentModePresetSchema,
+  type AgentModePreset,
+} from '@shared/schemas/agentPresets';
+import {
+  type AgentEntry,
+  createKey,
+  deduplicateByName,
+  getAgent,
+  getVisibleAgents,
+  getWorkflowAgents,
+  getToolUseAgents,
+  loadAgents,
+} from '@agent/index';
+import { SETTINGS_VIEW_COMMANDS } from '@common/webview';
+import { showLoggedErrorMessage } from '@common/errors';
+import { WorkspaceStateKey, GlobalStateKey, workspaceSM, globalSM } from '@common/state';
+import { agentDirectories } from '@frontend/agents/AgentDirectoryManager';
+import { AbsoluteFS } from '@utils/files';
+
+import type { SettingsHandlerContext } from './SettingsHandlerContext';
+
+// Type helper for extracting specific message types
+type MessageFor<C extends SettingsViewInboundMessage['command']> = Extract<
+  SettingsViewInboundMessage,
+  { command: C }
+>;
+
+function entryToSelectionItem(
+  entry: AgentEntry,
+  enabledKeys: string[] | undefined,
+): AgentSelectionItem {
+  const key = createKey(entry.source, entry.name);
+  // undefined = never configured → all enabled; [] = explicitly empty → none enabled
+  const enabled =
+    enabledKeys === undefined ||
+    enabledKeys.includes(key) ||
+    enabledKeys.includes(entry.name);
+  return {
+    name: entry.name,
+    source: entry.source,
+    category: entry.category,
+    description: entry.description,
+    hasPath: Boolean(entry.path),
+    filePath: entry.path || undefined,
+    tools: entry.tools,
+    hasMultiple: entry.isMultiple ?? Boolean(entry.multiplePath),
+    hasMultiplePath: Boolean(entry.multiplePath),
+    enabled,
+  };
+}
+
+export function buildAgentSelectionItems(): {
+  workflow: AgentSelectionItem[];
+  toolUse: AgentSelectionItem[];
+} {
+  // No default → undefined means "never configured" (all enabled)
+  const workflowEnabled = workspaceSM.get<string[]>(
+    WorkspaceStateKey.ENABLED_AGENTS,
+  );
+  const toolUseEnabled = workspaceSM.get<string[]>(
+    WorkspaceStateKey.ENABLED_TOOL_USE_AGENTS,
+  );
+
+  const workflow = getWorkflowAgents()
+    .map((e) => entryToSelectionItem(e, workflowEnabled))
+    .sort((a, b) => a.name.localeCompare(b.name));
+  const toolUse = getToolUseAgents()
+    .map((e) => entryToSelectionItem(e, toolUseEnabled))
+    .sort((a, b) => a.name.localeCompare(b.name));
+  return { workflow, toolUse };
+}
+
+/**
+ * Agent selection, directory, and mode preset handler delegate.
+ */
+export class AgentHandlers {
+  constructor(
+    private readonly ctx: SettingsHandlerContext,
+    private readonly refreshAfterAgentMutation: () => Promise<void>,
+  ) {}
+
+  // ── Agent selection data ──
+
+  async sendAgentSelectionData(webview: vscode.Webview): Promise<void> {
+    await loadAgents();
+    const { workflow, toolUse } = buildAgentSelectionItems();
+    await webview.postMessage({
+      command: SETTINGS_VIEW_COMMANDS.UPDATE_AGENT_SELECTION,
+      workflow,
+      toolUse,
+    });
+  }
+
+  // ── Agent selection handlers ──
+
+  async handleOpenAgentYaml(
+    data: MessageFor<typeof SETTINGS_VIEW_CMD.OPEN_AGENT_YAML>,
+  ): Promise<void> {
+    try {
+      const key = createKey(data.agentSource, data.agentName);
+      const entry = getAgent(key);
+      if (!entry) {
+        await vscode.window.showErrorMessage(
+          `Agent not found: ${data.agentName} (${data.agentSource})`,
+        );
+        return;
+      }
+
+      const agentPath =
+        data.variant === 'multiple' ? entry.multiplePath : entry.path;
+      if (!agentPath) {
+        await vscode.window.showErrorMessage(
+          `No ${data.variant} YAML path for agent: ${data.agentName}`,
+        );
+        return;
+      }
+
+      const doc = await vscode.workspace.openTextDocument(agentPath);
+      await vscode.window.showTextDocument(doc, { preview: false });
+    } catch (error) {
+      await showLoggedErrorMessage(
+        this.ctx.channel,
+        'Failed to open agent YAML file',
+        error,
+      );
+    }
+  }
+
+  async handleSetAgentEnabled(
+    data: MessageFor<typeof SETTINGS_VIEW_CMD.SET_AGENT_ENABLED>,
+  ): Promise<void> {
+    try {
+      const stateKey =
+        data.category === 'workflow'
+          ? WorkspaceStateKey.ENABLED_AGENTS
+          : WorkspaceStateKey.ENABLED_TOOL_USE_AGENTS;
+      // No default → undefined means "never configured" (all enabled)
+      const raw = workspaceSM.get<string[]>(stateKey);
+      const current = raw ?? [];
+      const key = createKey(data.agentSource, data.agentName);
+
+      let updated: string[];
+      if (data.enabled) {
+        // Add by source:name key if not already present (check both formats)
+        if (current.includes(key) || current.includes(data.agentName)) {
+          updated = current;
+        } else {
+          updated = [...current, key];
+        }
+      } else if (raw === undefined) {
+        // Never configured (undefined) = "all enabled". To disable one agent,
+        // seed the config with all OTHER agents: undefined → [all except this one].
+        const allAgents =
+          data.category === 'workflow'
+            ? getWorkflowAgents()
+            : getToolUseAgents();
+        updated = allAgents
+          .map((e) => createKey(e.source, e.name))
+          .filter((k) => k !== key);
+      } else {
+        // Remove both name and source:name formats.
+        // An empty result means "nothing enabled" (not "all enabled").
+        updated = current.filter(
+          (entry) => entry !== key && entry !== data.agentName,
+        );
+      }
+
+      await workspaceSM.update(stateKey, updated);
+
+      await this.refreshAfterAgentMutation();
+    } catch (error) {
+      await showLoggedErrorMessage(
+        this.ctx.channel,
+        'Failed to update agent visibility',
+        error,
+      );
+    }
+  }
+
+  async handleSetAllAgentsEnabled(
+    data: MessageFor<typeof SETTINGS_VIEW_CMD.SET_ALL_AGENTS_ENABLED>,
+  ): Promise<void> {
+    try {
+      const stateKey =
+        data.category === 'workflow'
+          ? WorkspaceStateKey.ENABLED_AGENTS
+          : WorkspaceStateKey.ENABLED_TOOL_USE_AGENTS;
+
+      const allAgents =
+        data.category === 'workflow' ? getWorkflowAgents() : getToolUseAgents();
+
+      const sourceAgents = allAgents.filter((e) => e.source === data.source);
+      const targetKeys = new Set(
+        sourceAgents.map((e) => createKey(e.source, e.name)),
+      );
+      // Legacy entries may use plain agent names without source prefix
+      const targetLegacyNames = new Set(sourceAgents.map((e) => e.name));
+
+      // Resolve current state: undefined means "all enabled" (never configured)
+      const raw = workspaceSM.get<string[]>(stateKey);
+      const current = raw ?? allAgents.map((e) => createKey(e.source, e.name));
+
+      let updated: string[];
+      if (data.enabled) {
+        // Add all keys from the target source (deduplicated)
+        const currentSet = new Set(current);
+        updated = [
+          ...current,
+          ...[...targetKeys].filter((k) => !currentSet.has(k)),
+        ];
+      } else {
+        // Remove both source:name and legacy plain-name entries for this source
+        updated = current.filter(
+          (k) => !targetKeys.has(k) && !targetLegacyNames.has(k),
+        );
+      }
+
+      await workspaceSM.update(stateKey, updated);
+
+      await this.refreshAfterAgentMutation();
+    } catch (error) {
+      await showLoggedErrorMessage(
+        this.ctx.channel,
+        'Failed to update agent visibility',
+        error,
+      );
+    }
+  }
+
+  async handleOpenAgentFolder(
+    data: MessageFor<typeof SETTINGS_VIEW_CMD.OPEN_AGENT_FOLDER>,
+  ): Promise<void> {
+    try {
+      const folderPath = await agentDirectories.getDirectory(data.folderType);
+      if (!folderPath) {
+        await vscode.window.showErrorMessage(
+          `No local directory for agent source: ${data.folderType}`,
+        );
+        return;
+      }
+      await vscode.commands.executeCommand(
+        'revealFileInOS',
+        vscode.Uri.file(folderPath),
+      );
+    } catch (error) {
+      await showLoggedErrorMessage(
+        this.ctx.channel,
+        'Failed to open agent folder',
+        error,
+      );
+    }
+  }
+
+  async handleRevealAgentFile(
+    data: MessageFor<typeof SETTINGS_VIEW_CMD.REVEAL_AGENT_FILE>,
+  ): Promise<void> {
+    try {
+      const key = createKey(data.agentSource, data.agentName);
+      const entry = getAgent(key);
+      if (!entry?.path) {
+        await vscode.window.showErrorMessage(
+          `Agent not found or has no file: ${data.agentName}`,
+        );
+        return;
+      }
+      await vscode.commands.executeCommand(
+        'revealFileInOS',
+        vscode.Uri.file(entry.path),
+      );
+    } catch (error) {
+      await showLoggedErrorMessage(
+        this.ctx.channel,
+        'Failed to reveal agent file',
+        error,
+      );
+    }
+  }
+
+  async handleCreateAgent(
+    data: MessageFor<typeof SETTINGS_VIEW_CMD.CREATE_AGENT>,
+  ): Promise<void> {
+    if (data.mode === 'template') {
+      await this.createAgentFromTemplate();
+    } else {
+      await vscode.commands.executeCommand(
+        'texra.createAgentWithAI',
+        data.category,
+      );
+    }
+
+    await this.refreshAfterAgentMutation();
+  }
+
+  async handleCustomizeAgent(
+    data: MessageFor<typeof SETTINGS_VIEW_CMD.CUSTOMIZE_AGENT>,
+  ): Promise<void> {
+    try {
+      const key = createKey(data.agentSource, data.agentName);
+      const entry = getAgent(key);
+      if (!entry?.path) {
+        await vscode.window.showErrorMessage(
+          `Agent not found or has no file: ${data.agentName}`,
+        );
+        return;
+      }
+
+      const customDir = await agentDirectories.custom();
+      const normalizedCustomDir = path.resolve(customDir);
+      const sourceDir = await agentDirectories.getDirectory(data.agentSource);
+      const relativePath = sourceDir
+        ? path.relative(sourceDir, entry.path)
+        : path.basename(entry.path);
+      const targetPath = path.join(customDir, relativePath);
+
+      // Guard: only write files under the custom agent directory
+      if (
+        !path.resolve(targetPath).startsWith(normalizedCustomDir + path.sep)
+      ) {
+        await vscode.window.showErrorMessage(
+          `Refusing to copy: target path escapes the custom agents directory.`,
+        );
+        return;
+      }
+
+      await AbsoluteFS.ensureDir(path.dirname(targetPath));
+
+      // Avoid overwriting an existing custom copy with user edits
+      if (await AbsoluteFS.exists(targetPath)) {
+        const overwrite = 'Overwrite';
+        const choice = await vscode.window.showWarningMessage(
+          `A custom copy already exists: ${path.basename(targetPath)}`,
+          { modal: true },
+          overwrite,
+        );
+        if (choice !== overwrite) return;
+      }
+
+      await AbsoluteFS.copy(entry.path, targetPath, { overwrite: true });
+
+      // Also copy the _multiple variant if it stays inside the custom directory
+      if (entry.multiplePath) {
+        const multipleRelative = sourceDir
+          ? path.relative(sourceDir, entry.multiplePath)
+          : path.basename(entry.multiplePath);
+        const multipleTarget = path.join(customDir, multipleRelative);
+        if (
+          path
+            .resolve(multipleTarget)
+            .startsWith(normalizedCustomDir + path.sep)
+        ) {
+          await AbsoluteFS.ensureDir(path.dirname(multipleTarget));
+          try {
+            await AbsoluteFS.copy(entry.multiplePath, multipleTarget, {
+              overwrite: true,
+            });
+          } catch {
+            // _multiple variant may not exist on disk even if registered
+          }
+        }
+      }
+
+      const doc = await vscode.workspace.openTextDocument(
+        vscode.Uri.file(targetPath),
+      );
+      await vscode.window.showTextDocument(doc, { preview: false });
+
+      void vscode.window.showInformationMessage(
+        `Created custom copy: ${path.basename(targetPath)}`,
+      );
+
+      await this.refreshAfterAgentMutation();
+    } catch (error) {
+      await showLoggedErrorMessage(
+        this.ctx.channel,
+        'Failed to create custom agent copy',
+        error,
+      );
+    }
+  }
+
+  async handleDeleteCustomAgent(
+    data: MessageFor<typeof SETTINGS_VIEW_CMD.DELETE_CUSTOM_AGENT>,
+  ): Promise<void> {
+    try {
+      const key = createKey('custom', data.agentName);
+      const entry = getAgent(key);
+      if (!entry?.path) {
+        await vscode.window.showErrorMessage(
+          `Custom agent not found: ${data.agentName}`,
+        );
+        return;
+      }
+
+      // Guard: only delete files under the custom agent directory
+      const customDir = await agentDirectories.custom();
+      const normalizedPath = path.resolve(entry.path);
+      const normalizedCustomDir = path.resolve(customDir);
+      if (!normalizedPath.startsWith(normalizedCustomDir + path.sep)) {
+        await vscode.window.showErrorMessage(
+          `Refusing to delete: file is not inside the custom agents directory.`,
+        );
+        return;
+      }
+
+      await AbsoluteFS.delete(entry.path, { recursive: false });
+
+      // Also delete the _multiple variant if it is inside the custom directory
+      if (entry.multiplePath) {
+        const normalizedMultiple = path.resolve(entry.multiplePath);
+        if (normalizedMultiple.startsWith(normalizedCustomDir + path.sep)) {
+          try {
+            await AbsoluteFS.delete(entry.multiplePath, { recursive: false });
+          } catch (err) {
+            // Only ignore FileNotFound; surface other errors
+            if (
+              !(err instanceof vscode.FileSystemError) ||
+              err.code !== 'FileNotFound'
+            ) {
+              throw err;
+            }
+          }
+        }
+      }
+
+      void vscode.window.showInformationMessage(
+        `Deleted custom agent: ${data.agentName}`,
+      );
+
+      await this.refreshAfterAgentMutation();
+    } catch (error) {
+      await showLoggedErrorMessage(
+        this.ctx.channel,
+        'Failed to delete custom agent',
+        error,
+      );
+    }
+  }
+
+  // ── Custom agent directory handlers ──
+
+  async sendCustomAgentDir(webview: vscode.Webview): Promise<void> {
+    const configuredPath = globalSM
+      .get<string>(GlobalStateKey.CUSTOM_AGENT_DIR, '')
+      .trim();
+    const isDefault = configuredPath === '';
+    const resolvedPath = await agentDirectories.custom();
+    await webview.postMessage({
+      command: SETTINGS_VIEW_COMMANDS.UPDATE_CUSTOM_AGENT_DIR,
+      path: resolvedPath,
+      isDefault,
+    });
+  }
+
+  async handleSetCustomAgentDir(): Promise<void> {
+    try {
+      const selectedPath = await agentDirectories.promptCustom();
+      if (!selectedPath) return;
+      await this.refreshAgentDirUI();
+    } catch (error) {
+      await showLoggedErrorMessage(
+        this.ctx.channel,
+        'Failed to set custom agent directory',
+        error,
+      );
+    }
+  }
+
+  async handleResetCustomAgentDir(): Promise<void> {
+    try {
+      await globalSM.update(GlobalStateKey.CUSTOM_AGENT_DIR, '');
+      await this.refreshAgentDirUI();
+    } catch (error) {
+      await showLoggedErrorMessage(
+        this.ctx.channel,
+        'Failed to reset custom agent directory',
+        error,
+      );
+    }
+  }
+
+  // ── Agent mode preset handlers ──
+
+  async sendAgentModePresets(webview: vscode.Webview): Promise<void> {
+    await webview.postMessage({
+      command: SETTINGS_VIEW_COMMANDS.UPDATE_AGENT_MODE_PRESETS,
+      customPresets: this.getCustomPresets(),
+    });
+  }
+
+  async handleApplyAgentModePreset(
+    data: MessageFor<typeof SETTINGS_VIEW_CMD.APPLY_AGENT_MODE_PRESET>,
+  ): Promise<void> {
+    try {
+      const preset =
+        AGENT_MODE_PRESETS.find((p) => p.id === data.presetId) ??
+        this.getCustomPresets().find((p) => p.id === data.presetId);
+      if (!preset) {
+        await vscode.window.showErrorMessage(
+          `Unknown preset: ${data.presetId}`,
+        );
+        return;
+      }
+
+      await loadAgents();
+
+      const workflowKeys = this.resolveAgentKeys(
+        'workflow',
+        new Set(preset.workflowAgents),
+      );
+      const toolUseKeys = this.resolveAgentKeys(
+        'toolUse',
+        new Set(preset.toolUseAgents),
+      );
+
+      await workspaceSM.update(WorkspaceStateKey.ENABLED_AGENTS, workflowKeys);
+      await workspaceSM.update(
+        WorkspaceStateKey.ENABLED_TOOL_USE_AGENTS,
+        toolUseKeys,
+      );
+
+      await this.refreshAfterAgentMutation();
+
+      void vscode.window.showInformationMessage(
+        `Applied "${preset.name}" preset`,
+      );
+    } catch (error) {
+      await showLoggedErrorMessage(
+        this.ctx.channel,
+        'Failed to apply agent mode preset',
+        error,
+      );
+    }
+  }
+
+  async handleSaveAgentModePreset(
+    _data: MessageFor<typeof SETTINGS_VIEW_CMD.SAVE_AGENT_MODE_PRESET>,
+  ): Promise<void> {
+    try {
+      const name = await vscode.window.showInputBox({
+        prompt: 'Name for the new preset',
+        placeHolder: 'e.g. My Research Setup',
+        validateInput: (v) => (v.trim() ? null : 'Name cannot be empty'),
+      });
+      if (!name) return; // cancelled
+
+      await loadAgents();
+
+      const workflowAgents = getVisibleAgents('workflow').map((e) => e.name);
+      const toolUseAgents = getVisibleAgents('toolUse').map((e) => e.name);
+
+      const preset: AgentModePreset = {
+        id: `custom-${Date.now()}`,
+        name: name.trim(),
+        description: `Custom preset: ${[...toolUseAgents, ...workflowAgents].join(', ')}`,
+        icon: 'codicon-bookmark',
+        workflowAgents,
+        toolUseAgents,
+      };
+
+      const existing = this.getCustomPresets();
+      await workspaceSM.update(WorkspaceStateKey.CUSTOM_AGENT_PRESETS, [
+        ...existing,
+        preset,
+      ]);
+
+      await this.ctx.withActiveWebview((w) => this.sendAgentModePresets(w));
+
+      void vscode.window.showInformationMessage(
+        `Saved preset "${name.trim()}"`,
+      );
+    } catch (error) {
+      await showLoggedErrorMessage(
+        this.ctx.channel,
+        'Failed to save agent mode preset',
+        error,
+      );
+    }
+  }
+
+  async handleDeleteAgentModePreset(
+    data: MessageFor<typeof SETTINGS_VIEW_CMD.DELETE_AGENT_MODE_PRESET>,
+  ): Promise<void> {
+    try {
+      const presets = this.getCustomPresets();
+      const target = presets.find((p) => p.id === data.presetId);
+      if (!target) return;
+
+      const confirm = await vscode.window.showWarningMessage(
+        `Delete preset "${target.name}"?`,
+        { modal: true },
+        'Delete',
+      );
+      if (confirm !== 'Delete') return;
+
+      const updated = presets.filter((p) => p.id !== data.presetId);
+      await workspaceSM.update(WorkspaceStateKey.CUSTOM_AGENT_PRESETS, updated);
+
+      await this.ctx.withActiveWebview((w) => this.sendAgentModePresets(w));
+    } catch (error) {
+      await showLoggedErrorMessage(
+        this.ctx.channel,
+        'Failed to delete agent mode preset',
+        error,
+      );
+    }
+  }
+
+  // ── Private helpers ──
+
+  private async createAgentFromTemplate(): Promise<void> {
+    try {
+      const name = await vscode.window.showInputBox({
+        prompt: 'Enter a name for the new agent (without .yaml extension)',
+        placeHolder: 'my_agent',
+        validateInput: (value) => {
+          if (!value) return 'Name cannot be empty';
+          if (value.includes('/') || value.includes('\\'))
+            return 'Name cannot contain path separators';
+          if (value.includes(' ')) return 'Use underscores instead of spaces';
+          if (/[:#\[\]{}|>&*!%@`]/.test(value))
+            return 'Name cannot contain YAML-special characters';
+          return null;
+        },
+      });
+      if (!name) return;
+
+      const customDir = await agentDirectories.custom();
+      await AbsoluteFS.ensureDir(customDir);
+
+      const fileName = name.endsWith('.yaml') ? name : `${name}.yaml`;
+      const filePath = path.join(customDir, fileName);
+
+      if (await AbsoluteFS.exists(filePath)) {
+        await vscode.window.showWarningMessage(
+          `A file named "${fileName}" already exists in the custom agents folder.`,
+        );
+        return;
+      }
+
+      const baseName = name.replace(/\.yaml$/, '');
+      const template = `# --- Agent Inheritance (Optional) ---
+# inherits: base
+
+name: ${baseName}
+
+# --- Agent Settings ---
+settings:
+  temperature: 0.1
+  isRewrite: true
+  documentTag: document
+  endTag: '</document>'
+  outputExt: tex
+  prefills:
+    - "<document>"
+
+# --- Agent Prompts ---
+prompts:
+  systemPrompt: |
+    [Define the AI's role and core instructions]
+
+  userPrefix: |
+    [Provide context using variables like {{ INPUT_CONTENT }}]
+
+  userRequest:
+    - |
+        [Define the initial task prompt]
+    - |
+        [Optional reflection prompt — remove this item if you only need one round]
+`;
+
+      await AbsoluteFS.write(filePath, template);
+      const doc = await vscode.workspace.openTextDocument(
+        vscode.Uri.file(filePath),
+      );
+      await vscode.window.showTextDocument(doc);
+    } catch (error) {
+      await showLoggedErrorMessage(
+        this.ctx.channel,
+        'Failed to create agent from template',
+        error,
+      );
+    }
+  }
+
+  /** Refresh agent dir + selection after a directory change. */
+  private async refreshAgentDirUI(): Promise<void> {
+    await agentDirectories.refreshAfterDirChange();
+    await this.ctx.withActiveWebview(async (w) => {
+      await Promise.all([
+        this.sendCustomAgentDir(w),
+        this.sendAgentSelectionData(w),
+      ]);
+    });
+    void vscode.commands.executeCommand('texra.refreshAllOptions');
+  }
+
+  /** Read and validate custom presets from workspace state. */
+  private getCustomPresets(): AgentModePreset[] {
+    const raw = workspaceSM.get(WorkspaceStateKey.CUSTOM_AGENT_PRESETS, []);
+    const parsed = z.array(AgentModePresetSchema).safeParse(raw);
+    return parsed.success ? parsed.data : [];
+  }
+
+  /** Resolve agent names to source:name keys for the given category. */
+  private resolveAgentKeys(
+    category: 'workflow' | 'toolUse',
+    names: Set<string>,
+  ): string[] {
+    const entries = deduplicateByName(
+      category === 'workflow' ? getWorkflowAgents() : getToolUseAgents(),
+    );
+    return entries
+      .filter((e) => names.has(e.name))
+      .map((e) => createKey(e.source, e.name));
+  }
+}
