+++ src/shared/schemas/subagentProgress.ts
@@ -1,0 +1,42 @@
+/**
+ * Subagent progress update types.
+ *
+ * These are pure data types with no implementation dependencies,
+ * used by both @agent/ and @tools/ to communicate subagent progress.
+ * Living in @shared/schemas/ breaks the circular dependency between
+ * agent runtime and tools.
+ */
+
+import type { TodoItem } from './todo';
+
+/** Todo state changed in a tool-use subagent. */
+export interface TodoProgressUpdate {
+  readonly kind: 'todos';
+  readonly todos: TodoItem[];
+}
+
+/** Workflow round completed. */
+export interface RoundProgressUpdate {
+  readonly kind: 'round';
+  readonly currentRound: number;
+  readonly totalRounds: number;
+}
+
+/** Periodic overview of tool-use subagent activity. */
+export interface OverviewProgressUpdate {
+  readonly kind: 'overview';
+  readonly toolCallCount: number;
+  readonly filesChanged: string[];
+  readonly cost?: number;
+}
+
+/** Subagent has finished initialization and is about to call the model. */
+export interface StartedProgressUpdate {
+  readonly kind: 'started';
+}
+
+export type SubagentProgressUpdate =
+  | TodoProgressUpdate
+  | RoundProgressUpdate
+  | OverviewProgressUpdate
+  | StartedProgressUpdate;
